WebConfig separated from AppConfig
