+++ src/main/java/com/filemanager/config/AppConfig.java
@@ -1,33 +1,17 @@
 package com.filemanager.config;
 
-import java.util.List;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.servlet.ViewResolver;
-import org.springframework.web.servlet.config.annotation.DefaultServletHandlerConfigurer;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 import org.springframework.web.servlet.view.JstlView;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.hibernate4.Hibernate4Module;
-
 @Configuration
-@EnableWebMvc
 @EnableTransactionManagement
-@ComponentScan(basePackages = { "com.filemanager.frontend","com.filemanager.backend"})
-public class AppConfig extends WebMvcConfigurerAdapter {
-
-	@Override
-	public void configureDefaultServletHandling(DefaultServletHandlerConfigurer configurer) {
-		configurer.enable();
-	}
+@ComponentScan(basePackages = { "com.filemanager.frontend", "com.filemanager.backend" })
+public class AppConfig {
 
 	@Bean
 	public ViewResolver jspViewResolver() {
@@ -36,25 +20,6 @@ public class AppConfig extends WebMvcConfigurerAdapter {
 		resolver.setPrefix("/");
 		resolver.setSuffix(".html");
 		return resolver;
-	}
-	
-	public MappingJackson2HttpMessageConverter jacksonMessageConverter(){
-	    MappingJackson2HttpMessageConverter messageConverter = new  MappingJackson2HttpMessageConverter();
-
-	    ObjectMapper mapper = new ObjectMapper();
-	    //Registering Hibernate4Module to support lazy objects
-	    mapper.registerModule(new Hibernate4Module());
-
-	    messageConverter.setObjectMapper(mapper);
-	    return messageConverter;
-
-	}
-
-	@Override
-	public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
-	    //Here we add our custom-configured HttpMessageConverter
-	    converters.add(jacksonMessageConverter());
-	    super.configureMessageConverters(converters);
 	}
 
 }
